refactor(mcp-conformance): extract Node-side watchdog + SDK-client teardown helper (rf2-sems4)
Three test/*.js files repeated ~40 LoC each of pure framing: watchdog
setTimeout/clearTimeout, main().then/catch, SDK Client +
StdioClientTransport construction, transport.stderr piping, and the
try/catch teardown shape. Adding a fourth Node test (causa-mcp impl,
a third pair2 variant) would have meant ~110 LoC of ceremony.

Extract `runWithWatchdog({ watchdogMs, transportSpec, clientName,
clientVersion }, async (client, transport) => { ... })` to
`tools/mcp-conformance/test/_runner.js`. The runner installs the
watchdog, constructs the transport (with stderr piping spliced in
automatically), constructs the SDK Client, runs `client.connect()`,
invokes the workflow body, and tears down the transport on every exit
path. Each test body shrinks to the workflow steps that actually
differ.

`live-pair2-overflow.js`'s SKIP branch stays as a pre-flight
`process.exit(0)` BEFORE the runner is invoked — we don't spawn a
child or install a watchdog when the test is skipping.
